+++ compiler/ext/config.ml
@@ -1,23 +1,35 @@
 let version = "4.06.1+BS"
 
-(* FIXME: Unreliable resolution *)
+(* This resolves the location of the standard library starting from the location of bsc.exe,
+  handling different supported package layouts. *)
 let standard_library =
-  let ( // ) = Filename.concat in
-  let exe_path = Sys.executable_name in
-  if Ext_string.contain_substring exe_path ("node_modules" // "@rescript") then
-    (* node_modules/@rescript/{platform}/bin *)
-    Filename.dirname exe_path // Filename.parent_dir_name
-    // Filename.parent_dir_name // Filename.parent_dir_name // "rescript"
-    // "lib" // "ocaml"
-  else if Ext_string.contain_substring exe_path ("node_modules" // "rescript")
-  then
-    (* node_modules/rescript/{platform} *)
-    Filename.dirname exe_path // Filename.parent_dir_name // "lib" // "ocaml"
-  else
-    (* git repo: rescript/packages/@rescript/{platform}/bin *)
-    Filename.dirname exe_path // Filename.parent_dir_name
-    // Filename.parent_dir_name // Filename.parent_dir_name
-    // Filename.parent_dir_name // "lib" // "ocaml"
+  let build_path rest path =
+    String.concat Filename.dir_sep (List.rev_append rest path)
+  in
+  match
+    Sys.executable_name |> Filename.dirname
+    |> String.split_on_char Filename.dir_sep.[0]
+    |> List.rev
+  with
+  (* 1. Packages installed via pnpm
+     - bin:    node_modules/.pnpm/@rescript+darwin-arm64@12.0.0-alpha.13/node_modules/@rescript/darwin-arm64/bin
+     - stdlib: node_modules/rescript/lib/ocaml (symlink)
+  *)
+  | "bin" :: _platform :: "@rescript" :: "node_modules" :: _package :: ".pnpm"
+    :: "node_modules" :: rest ->
+    build_path rest ["node_modules"; "rescript"; "lib"; "ocaml"]
+  (* 2. Packages installed via npm
+     - bin:    node_modules/@rescript/{platform}/bin
+     - stdlib: node_modules/rescript/lib/ocaml
+  *)
+  | "bin" :: _platform :: "@rescript" :: "node_modules" :: rest ->
+    build_path rest ["node_modules"; "rescript"; "lib"; "ocaml"]
+  (* 3. Several other cases that can occur in local development, e.g.
+     - bin:    <repo>/packages/@rescript/{platform}/bin, <repo>/_build/install/default/bin
+     - stdlib: <repo>/lib/ocaml
+  *)
+  | _ :: _ :: _ :: _ :: rest -> build_path rest ["lib"; "ocaml"]
+  | _ -> ""
 
 let standard_library_default = standard_library